% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_monotonic.R
\name{clean_monotonic}
\alias{clean_monotonic}
\title{Removes all non-monotonically increasing or decreasing data points for each
site (data with the same ID) in the Beck's Hybrids dataset.}
\usage{
clean_monotonic(orig_df)
}
\arguments{
\item{orig_df}{A data frame created from the Beck's dataset.}

\item{ID_list}{A list of all unique IDs within the Beck's dataset.}
}
\value{
A data frame of Beck's data w/o divergent data points.
}
\description{
**Will be further generalized to handle a range of datasets (with some
formatting requirements)
}
