% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_WLS.R
\name{clean_WLS}
\alias{clean_WLS}
\title{Filters data using a weighted least-squares linear regression model and a 
moving window (aims to minimize squared error).}
\usage{
clean_WLS(orig_df)
}
\arguments{
\item{orig_df}{A data frame created from the Beck's dataset.}

\item{ID_list}{A list of all unique IDs within the Beck's dataset.}
}
\value{
A data frame of Beck's data w/o divergent data points.
}
\description{
Filters data using a weighted least-squares linear regression model and a 
moving window (aims to minimize squared error).
}
