% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlog.R
\name{dlog}
\alias{dlog}
\title{Calculates RMSE values for the optimized double logistic function.}
\usage{
dlog(par, t, NDVI)
}
\arguments{
\item{par}{Parameters for prediction function (as listed above).}

\item{t}{Days since planting.}

\item{NDVI}{Actual observed NDVI for the site at a time t.}
}
\value{
The RMSE of the double-logistic function predictions.
}
\description{
I.e., evaluates the success of the double-logistic function at predicting
DOS.
}
\details{
#' The parameters include:
 1) mn: winter NDVI
 2) mx: maximum NDVI
 3) sos: start-of-season
 4) rsp: initial slope
 5) eos: end-of-season
 6) rau: ending slope
}
