% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_double_logistic.R
\name{fit_double_logistic}
\alias{fit_double_logistic}
\title{Iterates through all training plots, fits a double logistic, and records the
fitted parameters.}
\usage{
fit_double_logistic(x, t)
}
\arguments{
\item{x}{vector of NDVI values}

\item{t}{vector of DOY values with NDVI values}
}
\value{
A data frame of the fitted double logistic function for each site.
}
\description{
If the model is not able to be fit to the site data, the site is skipped.
}
