% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_normal.R
\name{fit_normal}
\alias{fit_normal}
\title{Iterates through all training plots, fits a normal dist, and records the
fitted parameters.}
\usage{
fit_normal(x, t)
}
\arguments{
\item{x}{vector of NDVI values}

\item{t}{vector of DOY values with NDVI values}
}
\value{
A data frame of the fitted normal function for each site.
}
\description{
If the model is not able to be fit to the site data, the site is skipped.
}
