% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met.R
\name{get_met}
\alias{get_met}
\title{Given a plot location (lat/lon) and year, calculates monthly precipitation
and cumulative GDD (Growing Degree Days).}
\usage{
get_met(lat, lon, year)
}
\arguments{
\item{lat}{plot latitude}

\item{lon}{plot longitude}

\item{year}{year that we need data for}
}
\value{
A data frame containing cumulative GDD and precipitation for each month of the year
}
\description{
Retrieves precipitation, temperature data from the meterological database
'ERA5 hourly data on single levels from 1979 to present'.
}
