% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_double_log.R
\name{optimize_double_log}
\alias{optimize_double_log}
\title{Fits a double logistic curve to observed values using a function described in
Beck et al. (2006) (equation 3).}
\usage{
optimize_double_log(
  x,
  t,
  tout = t,
  weight = TRUE,
  hessian = FALSE,
  plot = FALSE,
  ninit = 30
)
}
\arguments{
\item{x}{The vector (or time-series) to fit.}

\item{t}{Time steps.}

\item{tout}{Time steps of output (can be used for interpolation).}

\item{weight}{Weighting scheme (Beck), higher values get higher weight, etc.}

\item{hessian}{Hessian matrix used to compute standard error of parameters.}

\item{plot}{If true, the plot is iterating until best logistic fit.}

\item{ninit}{Number of initial params on which to optimize DLOG.}
}
\value{
A matrix of DOS predictions.
}
\description{
Fits a double logistic curve to observed values using a function described in
Beck et al. (2006) (equation 3).
}
