% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_DOS.R
\name{predict_DOS}
\alias{predict_DOS}
\title{Parse through all pixels and record optimized DOS.}
\usage{
predict_DOS(dat, doys, fA, fB, start_d, end_d)
}
\arguments{
\item{dat}{A x-y-n array of NDVI data.}

\item{doys}{A list of days since planting at each observed date.}

\item{fA}{Function for 1st Venancio equation.}

\item{fB}{Function for 2nd Venancio equation.}

\item{start_d}{The earliest DOS value for the observed region.}

\item{end_d}{The latest DOS value for the observed region.}
}
\value{
A matrix of DOS predictions.
}
\description{
Calculates optimized DOS by averaging the results of both Venancio
equations. Generates predictions for pixels with at least 3 dates of data.
}
\details{
Predictions outside of a reasonable DOS range (before/after insurance 
coverage dates) are set to -999.
}
