% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDLOG.R
\name{predictDLOG}
\alias{predictDLOG}
\title{Predict NDVI for any DOY based on the fitted DLOG parameters.}
\usage{
predictDLOG(params, t)
}
\arguments{
\item{params}{Parameters for prediction function (as listed above).}

\item{t}{Days since planting.}
}
\value{
Planting date predictions from input parameters, time after DOS.
}
\description{
The parameters include: 
 - mn: winter NDVI
 - mx: maximum NDVI
 - sos: start-of-season
 - rsp: initial slope
 - eos: end-of-season
 - rau: ending slope
}
