% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_3x3.R
\name{smooth_3x3}
\alias{smooth_3x3}
\title{Applies a uniform 3x3 smoothing window across the input matrix
(excluding pixels equal to NA or -999).}
\usage{
smooth_3x3(dos.mat)
}
\arguments{
\item{dos.mat}{The matrix of pixels to be smoothed.}
}
\value{
A smoothed matrix (excluding NA/-999 pixels).
}
\description{
A smaller smoothing window (e.g., 2x2, 2x3) is applied to edge and corner pixels.
}
