% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met.R
\name{get_met}
\alias{get_met}
\title{Retrieves hourly total precipitation, maximum temperature, minimum 
temperature, soil temperature, and soil moisture data from the CDS reanalysis
data product 'ERA5 hourly data on single levels from 1979 to present'.}
\usage{
get_met(lat, lon, date)
}
\arguments{
\item{lat}{Site latitude coordinate.}

\item{lon}{Site longitude coordinate.}

\item{date}{blaaah}
}
\value{
A data frame containing cumulative GDD and precipitation for each month of the year
}
\description{
Retrieves hourly total precipitation, maximum temperature, minimum 
temperature, soil temperature, and soil moisture data from the CDS reanalysis
data product 'ERA5 hourly data on single levels from 1979 to present'.
}
